Migrating to PSR-4 Namespaced Package (DariusIII\NetNntp) (#2)
* Performance overhaul, enum response codes, Composer autoload

BREAKING: Replace all NET_NNTP_PROTOCOL_RESPONSECODE_* defines with
ResponseCode enum. Remove deprecated v1.0/v1.1 API shims from
getOverview(), getArticle(), getHeader(), getBody(), post().

Performance:
- Move _clearOpensslErrors() out of _getTextResponse() loop (SSL/TLS).
- Use str_replace for dot-stuffing; drop redundant trim() in cmd*.
- cmdListNewsgroups/cmdXGTitle: regex → strpos/substr.
- cmdNewnews: return _getTextResponse() directly; merge fwrite in _sendArticle().
- Cache logger debug flag; use str_ends_with/str_starts_with; simplify cmd building.

Autoload:
- Remove manual require from NNTP/Client.php and NNTP/Protocol/Client.php;
  rely on Composer classmap. Benchmark loads via vendor/autoload.php.

Docs:
- Add README (credits: Heino, DariusIII, CicerBro), changelog v2.2.0,
  benchmark docblock (row meanings, legacy vs optimized map, variance note).

php benchmark_integration.php --host=news-eu.usenetexpress.com  --group=alt.binaries.boneless --user=une12172024 --pass=llMQi3JAk --range=137480081588-137480085088  --iterations=10   --include-legacy-reader
Connected to news-eu.usenetexpress.com:119 group=alt.binaries.boneless range=137480081588-137480085088
------------------------------------------------------------------------------------------------------------------------
getOverview() current              rows=3501    iter=10  median= 1070.726ms min=  527.811ms max= 2180.880ms avg= 1170.233ms
XOVER raw                          rows=3501    iter=10  median= 1005.436ms min=  529.523ms max= 1713.556ms avg= 1119.193ms
XOVER + legacy map                 rows=3501    iter=10  median=  960.003ms min=  555.216ms max= 1115.987ms avg=  882.083ms
XOVER + optimized map              rows=3501    iter=10  median=  843.918ms min=  606.183ms max= 1222.262ms avg=  882.525ms
XOVER using legacy reader          rows=3501    iter=10  median= 1351.050ms min=  874.969ms max= 2803.355ms avg= 1447.361ms
------------------------------------------------------------------------------------------------------------------------
Delta vs getOverview() current (median):
  getOverview() current                 +0.00%
  XOVER raw                             -6.10%
  XOVER + legacy map                   -10.34%
  XOVER + optimized map                -21.18%
  XOVER using legacy reader            +26.18%
------------------------------------------------------------------------------------------------------------------------

* Update README.md

* Migrating to PSR-4 Namespaced Package (DariusIII\NetNntp)

This release replaces the legacy PEAR-style class naming with PSR-4 namespaces under `DariusIII\NetNntp`. Source files have moved from `NNTP/` to `src/` and Composer autoloading has switched from `classmap` to `psr-4`.

Contains many breaking changes. See UPGRADE.md

* Remove deprecated v1.0 API and tighten types

Cleanup legacy v1.0 compatibility code and tighten parameter types. Changed $encryption from mixed to ?string in src/Client.php and src/Protocol/Client.php and removed the backwards-compat handling that treated an int $encryption as a port. Removed multiple deprecated methods and warnings (deprecated aliases and helpers such as quit, isConnected, getArticleRaw/getHeaderRaw/getBodyRaw/getNewNews/getReferencesOverview, setDebug, cmdXOver/cmdXROver extra-arg checks, and cmdXGTitle). Updated tests (tests/basics.php) to stop asserting against the removed isConnected() API. This simplifies the codebase and eliminates deprecated behavior; callers must pass explicit port values and use the current APIs instead of the old v1.0 shims.

* Docs: update README and UPGRADE for v4.0

Update documentation to reflect v4.0 API and namespace changes: switch README from classmap to PSR-4 usage, update example to use DariusIII\NetNntp\Client and Error, and qualify the ResponseCode enum path. Add an Upgrading section in README linking to UPGRADE.md and expand UPGRADE.md with a Deprecated Methods Removed table and guidance on the new connect() signature (use string encryption param). Also tidy credits and changelog references.

* Update README.md

* Ignore files

* Delete .phpunit.result.cache

* Remove benchmark script

Not useful anymore

* Fix typos
