+++ .gitignore
@@ -1,5 +1,6 @@
 # IDE
 .idea/
+.phpunit.result.cache
 
 # Composer
 /vendor/
