--- .travis.yml
@@ -1,14 +1,0 @@
-language: php
- 
-php:
-  - 5.4
-  - 5.5
-  - 5.6
-  - 7.0
-  - 7.1
- 
-before_script:
-  - composer self-update
-  - composer install --prefer-source --no-interaction --dev
- 
-script: phpunit
