+++ docs/examples/phpdoc/connect.php
@@ -1,4 +1,4 @@
 $posting = $nntp->connect('news.php.net');
-if (Net_NNTP_Error::isError($posting)) {
+if (\DariusIII\NetNntp\Error::isError($posting)) {
     // handle error
 }
