+++ docs/examples/phpdoc/getGroups.php
@@ -1,5 +1,5 @@
 $groups = $nntp->getGroups('*.pear.*');
-if (Net_NNTP_Error::isError($groupsummary)) {
+if (\DariusIII\NetNntp\Error::isError($groupsummary)) {
     // handle error
 }
