+++ phpunit.xml
@@ -1,19 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit
 	backupGlobals="false"
-	backupStaticAttributes="false"
-	
+
 	bootstrap="vendor/autoload.php"
-	
+
 	colors="true"
-	
-	convertErrorsToExceptions="true"
-	convertNoticesToExceptions="true"
-	convertWarningsToExceptions="true"
-	
+
 	processIsolation="false"
 	stopOnFailure="false"
-	syntaxCheck="false"
    >
 
 	<testsuites>
@@ -21,9 +15,5 @@
 			<directory suffix=".php">./tests/</directory>
 		</testsuite>
 	</testsuites>
-
-	<logging>
-		<log type="coverage-html" target="build/logs/html"/>
-	</logging>
 
 </phpunit>
